requirements.txt
Django
djangorestframework
mock
celery
